Compiling in Linux and process some complex stalled issues cases

+++ src/MEGASync/stalled_issues/gui/StalledIssueChooseWidget.cpp
@@ -46,7 +46,7 @@ void StalledIssueChooseWidget::setData(StalledIssueDataPtr data, const QString& 
         std::unique_ptr<mega::MegaNode> node(MegaSyncApp->getMegaApi()->getNodeByPath(mData->mPath.path.toStdString().c_str()));
         if(node)
         {
-            ui->fileSize->setText(Utilities::getSizeString(node->getSize()));
+            ui->fileSize->setText(Utilities::getSizeString((long long)node->getSize()));
         }
     }
     else
